geospatialscenario to geospatialscenario.py, tests
