+++ pvdeg/__init__.py
@@ -10,12 +10,10 @@ from . import degradation
 from . import design
 from . import fatigue
 from . import geospatial
-#from .geospatialscenario import GeospatialScenario
 from . import humidity
 from . import letid
 from . import montecarlo
 from . import pysam
-from .scenario import Scenario, GeospatialScenario
 from . import spectral
 from . import store
 from . import symbolic
@@ -24,6 +22,9 @@ from . import temperature
 from . import utilities
 from . import weather
 from . import diffusion
+
+from .scenario import Scenario
+from .geospatialscenario import GeospatialScenario
 
 __version__ = version("pvdeg")
