+++ pyproject.toml
@@ -73,7 +73,7 @@ docs = [
 test = [
     "pytest",
     "pytest-cov",
-    "sympy",
+    "scikit-learn",
 ]
 books = [
     "jupyter-book",
